api building updates
