+++ server/src/models/Order.ts
@@ -1,0 +1,7 @@
+
+export default class Order {
+    id: number;
+    userID: number;
+    quantity: number;
+    date: Date;
+}
